construct vm and binary_op

+++ include/value.hpp
@@ -2,7 +2,7 @@
 #include "common.hpp"
 #include <vector>
 namespace clox {
-    using Value = double;
+    using Value = int;
     using ValueArray = std::vector<Value>;
